+++ releasenotes.md
@@ -24,6 +24,7 @@
 - Implicit linking of libc math when libc math functions are used.
 - Allow even smaller memory limits.
 - Check unaligned array access.
+- Add "@structlike" for typedefs.
 
 ### Fixes
 - mkdir/rmdir would not work properly with substring paths on non-windows platforms.